tc: selftests: add helper script for virtualized testing
The vm.sh script bootstraps a kernel bzImage for tc testing and runs it on
a lightweight QEMU/KVM machine. The script accepts arbitrary commands to
run inside the VM, if given one, or runs 'tdc.py' (the whole test suite) by
default. The VM uses, by default, all CPU cores and 512M of the host.

It relies on virtme[1] for the QEMU/KVM bring up. It requires a minimum
KVM compatible kernel config or bzImage in order to boot. Sanitizers
and other debug/test features should be explicitly configured via Kconfig
or given in a custom config. The script only bootstraps settings relevant
to tdc when it bootstraps a kernel bzImage.

Examples:

- Bootstrap the kernel and trigger a kernel oops:
   ./vm.sh -- "echo c > /proc/sysrq-trigger"

- Generate a kernel config (if needed), bootstrap the kernel and
   run the Qdisc tdc tests:
   ./vm.sh -g -- ./tdc.py -c qdisc

- Bootstrap the kernel and run an interactive shell:
   ./vm.sh -s

- Dry-run and print the QEMU command line:
   ./vm.sh -d

- Use a specific kernel config file and constraint compilation to 2 CPUs:
   ./vm.sh -f ~/myconfig -j 2 -- "echo foo"

- Constraint the VM resources and run the default command:
   ./vm.sh -c 4 -m 512M

- Use a pre-compiled kernel image and run an interactive shell:
   ./vm.sh -i ~/mybzImage -s

[1] https://git.kernel.org/pub/scm/utils/kernel/virtme/virtme.git/

Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>