p4tc: add table class create, update, delete, get, flush and dump
This commit allows users to create, update, delete, get, flush and dump
table class.

It's important to note that write operations, such as create, update
and delete, can only be made if the pipeline is not sealed.

Per the P4 specification, table classes require the name of the control block.

As an example, if one were to create a table class named table1 in a
pipeline named myprog1, on control block "mycontrol", one would use
the following command:

tc p4template create tclass/myprog1/mycontrol/table1 tbcid 1 \
          keysz 32 nummasks 8 tentries 8192 \
          keys default action metact cmd set key hdrfield/1/parser5/1 \
          keys action metact cmd set key hdrfield/1/parser5/2 \
          postactions \
                action metact cmd beq results/hit true control pipe / jump 1 \
                              cmd act results/action control ok \
                              cmd act kernel.drop

It says this table will have two keys which is set by a metact
that assigns it to a specific header value extracted by the parser
(parser5). Note that we are creating keys without assigning a key ID.
That is because if we don't specify one, the kernel will give us one.
In the above example, the key ID will be 1 and 2 respectively, because it's
the first key we are creating for this table class. Also note that we use
the keyword "default" when defining the first key, this is to indicate that
this will be the default key for this table class. That is, if the user doesn't
specify a key when binding a table instance of this type to a filter,
the code will assume the key is the default one. If we have more than one key
and the user doesn't use the keyword default when defining any of them,
we assume the default key is the first one. The postaction specifies what
to execute after a match attempt has been performed in an instance of
this table class. In this particular example, if the match succeeded
we'll execute a metact action.

Some parameters such as tcount (number of table instances allowed for the
table class), nummasks (number of masks this table may have) and tentries
(maximum number of entries this table class may have) may be ommited. In
this case the default values are used, which are, respectively, 1, 8 and
256. There is also an optional argument (preactions), which will be
executed before the table key actions.

It's also important to note that the user can also specify table class ID
when creating a new table class. For example:

tc p4template create tclass/myprog1/mycontrol/table1 tbcid 1 keysz 16 ...

The table class ID is specified by using the "tbcid" keyword. However,
please note that if you don't specify a table class ID when creating, the
kernel will assign one for you.

If one were to update a table class named table1 in a pipeline named
myprog1, one would use the following command:

tc p4template update tclass/myprog1/mycontrol/table1 keys action gact index 4

Note that, in the above command, we are not specifying an ID for the
key. In that scenario, the command will create a new key for this table
type, instead of updating, and will assign a new ID for the key.

If one were to delete a table class named table1 pipeline named
myprog1, one would use the following command:

tc p4template del tclass/myprog1/mycontrol/table1

If one were to flush all the table class from a pipeline named myprog1,
one would use the following command:

tc p4template del tclass/myprog1/mycontrol/

If one were to get a table class named table1 from a pipeline named
myprog1, one would use the following command:

tc p4template get tclass/myprog1/mycontrol/table1

If one were to dump all the table class from a pipeline named myprog1, one
would use the following command:

tc p4template get tclass/myprog1/mycontrol/

Co-developed-by: Jamal Hadi Salim <jhs@mojatatu.com>
Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>