netns: Add ops init_id and exit_batch_id to pernet_operations
Add new init and exit_batch ops passing id as an argument.

Since most use cases of pernet operations rely on one ID that is static
per file or module, there is no need to pass the ID as an argument during
init or exit_batch, since we'll have this ID available locally in the
file or module where it was declared when these callback are invoked.
However, if our use case uses more than one ID per file or module, we
are unable to find out which ID the callback belongs to. In that case,
we need the callback to specify to which ID it relates to.

XXX: Remove this when rebasing to net-next. This was supersede in
upstream.

Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>