net/sched: act_api: Change act_base into an IDR
Change act_base from a list to an IDR.

With the introduction of action templates, we are allowing the creation
of actions on the fly. This means that we'll have use cases where the
action IDs need to be dynamically and not statically defined, as was the
case previously. Dynamic IDs are much harder to manage and thus it will
be simpler to change act_base to an IDR, instead of maintaining it as a
list and having to reimplement all of the ID managing features the IDR
code already implements.

Co-developed-by: Jamal Hadi Salim <jhs@mojatatu.com>
Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>