net/sched: act_api: Add init_ops to struct tc_action_op
As with all TC actions, dynamic actions possess a pernet structure which
holds the IDR where the action instances are stored. In the regular
actions (such as gact, mirred and ife), the net ID necessary to access this
pernet structure is static to the file where they are defined, which
means this ID is available to every function declared in the file. In
the case of dynamic functions, this ID can't be static to the file,
because we'll need a new ID for each new dynamic function. With that in
mind, we needed to find a way to pass this ID when needed, such as when the
init or the lookup callbacks are executed. To solve this issue, we
added the netid to struct tc_action_op and two new callbacks to the
struct tc_action_ops: init_ops and lookup_ops, which have, respectively,
the same prototype as init and lookup, but with the op as an extra
parameter. Also we needed to add a new refcount for dynamic functions
called dyn_ref.

XXX: This commit should disappear completely

Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>