p4tc: Add header field create, get, delete, flush and dump
This commit allows control to create, get, delete, flush and dump header
field objects. The created header fields are retrieved at runtime by
the parser. From a control plane interaction, a header field can only be
created once the appropriate parser is instantiated. At runtime, existing
header fields can be referenced for computation reasons from metact:
metact will use header fields to either create lookup keys or edit the
header fields.

Header fields are part of a pipeline and a parser instance and header
fields can only be created in an unsealed pipeline.

To create a header field, the user must issue the equivalent of the
following command:

tc p4template create hdrfield/ptables/myparser/ipv4/totlen

where ptables is the name of a pipeline, myparser is a name of a parser
instance, ipv4 is the name of header and totlen is a header field.
Essential header field information, such as length, position and type
are retrieved by consulting a JSON introspection file that will be
generated by the P4 compiler.

To delete a header field, the user must issue the equivalent of the
following command:

tc p4template del hdrfield/ptables/myparser/ipv4/totlen

where ptables is the name of pipeline, myparser is a name of a parser
instance, ipv4 is the name of header and totlen is a header field.

To retrieve meta-information from a header field, such as length,
position and type, the user must issue the equivalent of the following
command:

tc p4template get hdrfield/ptables/myparser/ipv4/totlen

where ptables is the name of pipeline, myparser is a name of a parser
instance, ipv4 is the name of header and totlen is a header field.

The user can also dump all the header fields available in a parser
instance using the equivalent of the following command:

tc p4template get hdrfield/ptables/myparser/

With that, the user will get all the header field names available in a
specific parser instance.

The user can also flush all the header fields available in a parser
instance using the equivalent of the following command:

tc p4template del hdrfield/ptables/myparser/

Header fields do not support update operations.

Co-developed-by: Jamal Hadi Salim <jhs@mojatatu.com>
Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>